requirements.txt
django==5.2.5
djangorestframework==3.16.1
pytest-django==4.11.1
faker==37.12.0
openpyxl==3.1.5
pyotp==2.9.0
django==5.2.5
djangorestframework
django-filter
django-cors-headers
Pillow
qrcode
openpyxl
